@@ -1,12 +1,20 @@
 name: video_player_example
 description: Demonstrates how to use the video_player plugin.
-version: 0.0.1
 publish_to: none
+
+environment:
+  sdk: ">=2.12.0 <3.0.0"
+  flutter: ">=1.12.13+hotfix.5"
 
 dependencies:
   flutter:
     sdk: flutter
   video_player:
+    # When depending on this package from a real application you should use:
+    #   video_player: ^x.y.z
+    # See https://dart.dev/tools/pub/dependencies#version-constraints
+    # The example app is bundled with the plugin so we use a path dependency on
+    # the parent directory to use the current plugin's version.
     path: ../
 
 dev_dependencies:
@@ -14,17 +22,14 @@ dev_dependencies:
     sdk: flutter
   flutter_driver:
     sdk: flutter
-  # integration_test:
-  # path: ../../../integration_test
-  pedantic: ^1.8.0
+  integration_test:
+    sdk: flutter
+  test: any
+  pedantic: ^1.10.0
 
 flutter:
   uses-material-design: true
   assets:
-    - assets/flutter-mark-square-64.png
-    - assets/Butterfly-209.mp4
-    - assets/bumble_bee_captions.srt
-
-environment:
-  sdk: ">=2.8.0 <3.0.0"
-  flutter: ">=1.12.13+hotfix.5 <2.0.0"
+   - assets/flutter-mark-square-64.png
+   - assets/Butterfly-209.mp4
+   - assets/bumble_bee_captions.srt
