@@ -1,4 +1,8 @@
-// Autogenerated from Pigeon (v0.1.7), do not edit directly.
+// Copyright 2013 The Flutter Authors. All rights reserved.
+// Use of this source code is governed by a BSD-style license that can be
+// found in the LICENSE file.
+
+// Autogenerated from Pigeon (v0.1.21), do not edit directly.
 // See also: https://pub.dev/packages/pigeon
 
 package io.flutter.plugins.videoplayer;
@@ -83,12 +87,23 @@ public class Messages {
       this.formatHint = setterArg;
     }
 
+    private HashMap httpHeaders;
+
+    public HashMap getHttpHeaders() {
+      return httpHeaders;
+    }
+
+    public void setHttpHeaders(HashMap setterArg) {
+      this.httpHeaders = setterArg;
+    }
+
     HashMap toMap() {
       HashMap<String, Object> toMapResult = new HashMap<>();
       toMapResult.put("asset", asset);
       toMapResult.put("uri", uri);
       toMapResult.put("packageName", packageName);
       toMapResult.put("formatHint", formatHint);
+      toMapResult.put("httpHeaders", httpHeaders);
       return toMapResult;
     }
 
@@ -102,6 +117,8 @@ public class Messages {
       fromMapResult.packageName = (String) packageName;
       Object formatHint = map.get("formatHint");
       fromMapResult.formatHint = (String) formatHint;
+      Object httpHeaders = map.get("httpHeaders");
+      fromMapResult.httpHeaders = (HashMap) httpHeaders;
       return fromMapResult;
     }
   }
@@ -597,7 +614,7 @@ public class Messages {
   private static HashMap wrapError(Exception exception) {
     HashMap<String, Object> errorMap = new HashMap<>();
     errorMap.put("message", exception.toString());
-    errorMap.put("code", null);
+    errorMap.put("code", exception.getClass().getSimpleName());
     errorMap.put("details", null);
     return errorMap;
   }
