@@ -1,4 +1,8 @@
-// Autogenerated from Pigeon (v0.1.7), do not edit directly.
+// Copyright 2013 The Flutter Authors. All rights reserved.
+// Use of this source code is governed by a BSD-style license that can be
+// found in the LICENSE file.
+
+// Autogenerated from Pigeon (v0.1.21), do not edit directly.
 // See also: https://pub.dev/packages/pigeon
 #import "messages.h"
 #import <Flutter/Flutter.h>
@@ -7,17 +11,19 @@
 #error File requires ARC to be enabled.
 #endif
 
-static NSDictionary *wrapResult(NSDictionary *result, FlutterError *error) {
+static NSDictionary<NSString *, id> *wrapResult(NSDictionary *result, FlutterError *error) {
   NSDictionary *errorDict = (NSDictionary *)[NSNull null];
   if (error) {
-    errorDict = [NSDictionary
-        dictionaryWithObjectsAndKeys:(error.code ? error.code : [NSNull null]), @"code",
-                                     (error.message ? error.message : [NSNull null]), @"message",
-                                     (error.details ? error.details : [NSNull null]), @"details",
-                                     nil];
+    errorDict = @{
+      @"code" : (error.code ? error.code : [NSNull null]),
+      @"message" : (error.message ? error.message : [NSNull null]),
+      @"details" : (error.details ? error.details : [NSNull null]),
+    };
   }
-  return [NSDictionary dictionaryWithObjectsAndKeys:(result ? result : [NSNull null]), @"result",
-                                                    errorDict, @"error", nil];
+  return @{
+    @"result" : (result ? result : [NSNull null]),
+    @"error" : errorDict,
+  };
 }
 
 @interface FLTTextureMessage ()
@@ -84,6 +90,10 @@ static NSDictionary *wrapResult(NSDictionary *result, FlutterError *error) {
   if ((NSNull *)result.formatHint == [NSNull null]) {
     result.formatHint = nil;
   }
+  result.httpHeaders = dict[@"httpHeaders"];
+  if ((NSNull *)result.httpHeaders == [NSNull null]) {
+    result.httpHeaders = nil;
+  }
   return result;
 }
 - (NSDictionary *)toMap {
@@ -93,7 +103,9 @@ static NSDictionary *wrapResult(NSDictionary *result, FlutterError *error) {
                                    (self.packageName ? self.packageName : [NSNull null]),
                                    @"packageName",
                                    (self.formatHint ? self.formatHint : [NSNull null]),
-                                   @"formatHint", nil];
+                                   @"formatHint",
+                                   (self.httpHeaders ? self.httpHeaders : [NSNull null]),
+                                   @"httpHeaders", nil];
 }
 @end
 
@@ -220,8 +232,8 @@ void FLTVideoPlayerApiSetup(id<FlutterBinaryMessenger> binaryMessenger, id<FLTVi
                binaryMessenger:binaryMessenger];
     if (api) {
       [channel setMessageHandler:^(id _Nullable message, FlutterReply callback) {
-        FlutterError *error;
         FLTCreateMessage *input = [FLTCreateMessage fromMap:message];
+        FlutterError *error;
         FLTTextureMessage *output = [api create:input error:&error];
         callback(wrapResult([output toMap], error));
       }];
@@ -235,8 +247,8 @@ void FLTVideoPlayerApiSetup(id<FlutterBinaryMessenger> binaryMessenger, id<FLTVi
                binaryMessenger:binaryMessenger];
     if (api) {
       [channel setMessageHandler:^(id _Nullable message, FlutterReply callback) {
-        FlutterError *error;
         FLTTextureMessage *input = [FLTTextureMessage fromMap:message];
+        FlutterError *error;
         [api dispose:input error:&error];
         callback(wrapResult(nil, error));
       }];
@@ -250,8 +262,8 @@ void FLTVideoPlayerApiSetup(id<FlutterBinaryMessenger> binaryMessenger, id<FLTVi
                binaryMessenger:binaryMessenger];
     if (api) {
       [channel setMessageHandler:^(id _Nullable message, FlutterReply callback) {
-        FlutterError *error;
         FLTLoopingMessage *input = [FLTLoopingMessage fromMap:message];
+        FlutterError *error;
         [api setLooping:input error:&error];
         callback(wrapResult(nil, error));
       }];
@@ -265,8 +277,8 @@ void FLTVideoPlayerApiSetup(id<FlutterBinaryMessenger> binaryMessenger, id<FLTVi
                binaryMessenger:binaryMessenger];
     if (api) {
       [channel setMessageHandler:^(id _Nullable message, FlutterReply callback) {
-        FlutterError *error;
         FLTVolumeMessage *input = [FLTVolumeMessage fromMap:message];
+        FlutterError *error;
         [api setVolume:input error:&error];
         callback(wrapResult(nil, error));
       }];
@@ -280,8 +292,8 @@ void FLTVideoPlayerApiSetup(id<FlutterBinaryMessenger> binaryMessenger, id<FLTVi
                binaryMessenger:binaryMessenger];
     if (api) {
       [channel setMessageHandler:^(id _Nullable message, FlutterReply callback) {
-        FlutterError *error;
         FLTPlaybackSpeedMessage *input = [FLTPlaybackSpeedMessage fromMap:message];
+        FlutterError *error;
         [api setPlaybackSpeed:input error:&error];
         callback(wrapResult(nil, error));
       }];
@@ -295,8 +307,8 @@ void FLTVideoPlayerApiSetup(id<FlutterBinaryMessenger> binaryMessenger, id<FLTVi
                                            binaryMessenger:binaryMessenger];
     if (api) {
       [channel setMessageHandler:^(id _Nullable message, FlutterReply callback) {
-        FlutterError *error;
         FLTTextureMessage *input = [FLTTextureMessage fromMap:message];
+        FlutterError *error;
         [api play:input error:&error];
         callback(wrapResult(nil, error));
       }];
@@ -310,8 +322,8 @@ void FLTVideoPlayerApiSetup(id<FlutterBinaryMessenger> binaryMessenger, id<FLTVi
                binaryMessenger:binaryMessenger];
     if (api) {
       [channel setMessageHandler:^(id _Nullable message, FlutterReply callback) {
-        FlutterError *error;
         FLTTextureMessage *input = [FLTTextureMessage fromMap:message];
+        FlutterError *error;
         FLTPositionMessage *output = [api position:input error:&error];
         callback(wrapResult([output toMap], error));
       }];
@@ -325,8 +337,8 @@ void FLTVideoPlayerApiSetup(id<FlutterBinaryMessenger> binaryMessenger, id<FLTVi
                binaryMessenger:binaryMessenger];
     if (api) {
       [channel setMessageHandler:^(id _Nullable message, FlutterReply callback) {
-        FlutterError *error;
         FLTPositionMessage *input = [FLTPositionMessage fromMap:message];
+        FlutterError *error;
         [api seekTo:input error:&error];
         callback(wrapResult(nil, error));
       }];
@@ -340,8 +352,8 @@ void FLTVideoPlayerApiSetup(id<FlutterBinaryMessenger> binaryMessenger, id<FLTVi
                binaryMessenger:binaryMessenger];
     if (api) {
       [channel setMessageHandler:^(id _Nullable message, FlutterReply callback) {
-        FlutterError *error;
         FLTTextureMessage *input = [FLTTextureMessage fromMap:message];
+        FlutterError *error;
         [api pause:input error:&error];
         callback(wrapResult(nil, error));
       }];
@@ -355,8 +367,8 @@ void FLTVideoPlayerApiSetup(id<FlutterBinaryMessenger> binaryMessenger, id<FLTVi
                binaryMessenger:binaryMessenger];
     if (api) {
       [channel setMessageHandler:^(id _Nullable message, FlutterReply callback) {
-        FlutterError *error;
         FLTMixWithOthersMessage *input = [FLTMixWithOthersMessage fromMap:message];
+        FlutterError *error;
         [api setMixWithOthers:input error:&error];
         callback(wrapResult(nil, error));
       }];
