@@ -1,4 +1,4 @@
-// Copyright 2020 The Chromium Authors. All rights reserved.
+// Copyright 2013 The Flutter Authors. All rights reserved.
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
@@ -31,7 +31,12 @@ class Caption {
   ///
   /// This is not recommended for direct use unless you are writing a parser for
   /// a new closed captioning file type.
-  const Caption({this.number, this.start, this.end, this.text});
+  const Caption({
+    required this.number,
+    required this.start,
+    required this.end,
+    required this.text,
+  });
 
   /// The number that this caption was assigned.
   final int number;
@@ -45,6 +50,15 @@ class Caption {
   /// The actual text that should appear on screen to be read between [start]
   /// and [end].
   final String text;
+
+  /// A no caption object. This is a caption with [start] and [end] durations of zero,
+  /// and an empty [text] string.
+  static const Caption none = Caption(
+    number: 0,
+    start: Duration.zero,
+    end: Duration.zero,
+    text: '',
+  );
 
   @override
   String toString() {
